Edits taken from camera cart computer using USB stick
I had to make a lot of configuration changes to fit the camera cart computer and these changes reflect that. Additionally, I needed to run through with a debugger and change some data processing issues. This push has incorrect rotations for the ToF sensor, but the data is still collected.